Refactor `native_impl`: replace `f64` with `f32` in `get_transform_translation`

+++ crates/canvas/src/canvas/native_impl.rs
@@ -264,8 +264,8 @@ impl CanvasRendering for CupidCanvas {
     }
 
     #[inline]
-    fn get_transform_translation(&self) -> (f64, f64) {
+    fn get_transform_translation(&self) -> (f32, f32) {
         let (tx, ty) = CupidCanvas::get_transform_translation(self);
-        (tx as f64, ty as f64)
+        (tx, ty)
     }
 }
